file rename & minor changes
The file name of the script is now “Skyparistic.scpt” which helps you
to better find it with launchbar or Alfred.

The “Do not Disturb” mode is now deactivated at the beginning block of
the “after call” commands, so the Notification at the End is more
likely to be seen.